*** empty log message ***

+++ unit_tests.cpp
@@ -100,7 +100,7 @@ BOOST_AUTO_TEST_CASE(tree_node_size)
 BOOST_AUTO_TEST_CASE(tree_minmax)
 {
   TreeNode root;
-  int8_t depth = 14;
+  int depth = 14;
   root.minmax(Board::WHITE, depth);
   std::cout << "\nDepth : " << depth << "\n"
 	    << root
